requirements.txt
numpy==1.24.4
Pillow==11.1.0
transformers==4.47.1
onnxruntime==1.20.1